requirements.txt
svgwrite>=1.4.3
pypinyin>=0.53.0
quart>=0.19.0
python-dotenv>=1.0.0
openai>=1.0.0
zhipuai>=2.0.0
dashscope>=1.13.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
